MOD: Change Magic Number 90, 120, 150 to CANMUNSMA, CANMUNMID, CANMUNLAR. These are the number of candlesticks.
Fix: Resolved the issue where TitleBarButton and TitleBarMenu were
displayed multiple times in multi mode. Changed the location using a
modifier from StockView to ContentView.

TODO: Display the stock code of the chart that is in focus on the title
bar. What is the method to notify ContentView of the currently focused
View?